Patched signed -> unsigned warning (issue 98)

+++ src/stream.cpp
@@ -400,8 +400,7 @@ namespace YAML
 		if (m_nPrefetchedUsed >= m_nPrefetchedAvailable)
 		{
 			std::streambuf *pBuf = m_input.rdbuf();
-			m_nPrefetchedAvailable = pBuf->sgetn(ReadBuffer(m_pPrefetched), 
-				YAML_PREFETCH_SIZE);
+			m_nPrefetchedAvailable = static_cast<std::size_t>(pBuf->sgetn(ReadBuffer(m_pPrefetched), YAML_PREFETCH_SIZE));
 			m_nPrefetchedUsed = 0;
 			if (!m_nPrefetchedAvailable)
 			{
